grai-integrations/source-cube/pyproject.toml
[tool.poetry]
name = "grai_source_cube"
version = "0.0.1"
description = ""
authors = ["Grai <dev@grai.io>"]
license = "Elastic-2.0"
packages = [
    { include = "grai_source_cube", from = "src" },
]
readme = "README.md"
homepage = "https://www.grai.io/"
repository = "https://github.com/grai-io/grai-core/tree/master/grai-integrations/source-cube"
documentation = "https://docs.grai.io/"

[tool.poetry.dependencies]
python = "^3.9.13"
pydantic = "^1.9.1"
grai-schemas = "^0.2.6"
multimethod = "^1.8, !=1.11"
python-dotenv = "^0.21.1"
pyjwt = "^2.8.0"
polyfactory = "^2.14.1"

[tool.poetry.group.dev.dependencies]
mypy = "^0.971"
pytest = "^7.2.0"
pre-commit = "^3.0.4"
types-requests = "^2.28.11.7"
faker = "^17.0.0"
grai-client = "^0.3.2"
ruff = "^0.2.2"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "I", "F"]

[tool.ruff.lint.isort]
known-first-party = ["grai_source_cube"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry_bumpversion.file."src/grai_source_cube/__init__.py"]
